Add navigation context and remove impossible tests
It doesn't seem like the testing-library makes allowances for navigation
elements so I'll have to just keep the tests to the level below where
global context items are required. It seems like a lot of the tests have
been removed now which might be a problem, but from memeory these were
all mostly testing the interoperability of the components. That is a
good thing to test, but we have simplified the mechanisms by which they
interact in the first place which might mean that these particular tests
are less important now.