Add function to derive props from harpstrata
This will be required by each type of the generic button being created.
Each will need to get the base props and then add a new prop in to it
depending on what is being changed. Each type of button will change a
different thing. Pozition  buttons will change Pozition for example.