One naive formulation of a getCovariantMap function
Another appraoch occurs to me which might not result in the necessary
creation of more types. It would be good to reuse the harpstrata types
as much as possible. To do that I think I should split apart the
funcitonality which gets multiple covariant groups from the
functionality which handles the identification of locked and variable
control vars.