Add another lower level conversion function
I'm searching for the right level of abstraction to build this
functionality out from. This seems like it could be useful, if very
verbose and apparently repetative. The repetition is difficult to remove
though since it's not obvious to the type system when we have a harp key
vs a root pitch and when we should accompany the types with pitch or
pozition ids.