Create empty CovarianceButtonList component
This component will present a list of CovarianceButtons, the types of
which will be governed by which of the CovariantTypes is locked.
The decision about which type is locked and which type we are ordering
by will be controlled by state somewhere, but initially we'll just set
those things statically.