Alter test due to flawed dependency
The ordering of the PozitionIds is not satisfactory at the moment since
they are ordered by root offset rather than by name. The ordering by
offset will emerge when either of the pitch types are the variedType so
thats unnecesary and will be fixed. When it gets fixed I don't want to
have to rewrite this test.