Add function to get varied list of control vars
This builds on the one to get a single control vars and will allow us to
build a CovariantsGroup list for a given covariants control set.

The covariants control set (working title) needs defining, but it's the
definition that the user will put in the UI. They will define one
covariant to be locked, and another to order by. The third will be
deduced and it's those last two which will be displayed in a grid
together so the user can anticipate how they will vary together.