Add functionality for remaining button types
And add test for button interactivity.

We don't have test for all of these permutations of this file at the
moment. That's partly because I don't want to have to maintain that many
similar tests although I accept that this means an important part of the
functionality is untested now which is a problem.

Another reason though is that I haven't finished designing all of
thisand if I'm going to add all of those tests, I'd rather add them at
the end when the design is more set, so that I don't need to constantly
maintain them during design change.