Rename getContolVars function
Control of what? Obviously they are vars.. So getConvariantControllers
is better. It's still helpful to have some prior knowledge here
unfortunately. Understanding that the word "Covariant" is being used to
refer to the subject around instances of single a CovariantSet and the
control of one variable within it is helpful to guide the reader to
understanding that this is going to return a CovariantControllers
object rather than anything in the area of CovarianceSeries which we
will soon be defining.