Rename CovariantsGroupList module to CovarianceSeries
Straight swap of one name for another. Just like the other renamings
that have been happening over the last hour, this name packs lots more
information in to a similar space. It's more domain specific which might
mean it requires some research to make sense of, but honestly, how
useful was knowing that something was a ...List when you don't know what
a CovariantsGroup is anyway. Research was always required.