Add a few more sample tests to be prudent
It would be unrealistic to test all of the possible input types at this
level of the functionality. Testing the inputs as their parent
CovariancePrimer types makes sense and testing just a sample of the
inputs and outputs also makes sense as long as there is one tests
checking the complete returned object.