Make hole number row non invasive annotation
The hole number row should not interrupt the flow of the rest of the
harp face. Having it have a 0 flex size in the harp face rows and giving
a z index to float above the other rows so that it's contents can be
seen seems like a sensible way to achieve this.