Make simple use of theme size on harp face
The logic of multiplying column rows by these themed sizes is that I
think a relationship between the cell size and the column size is going
to be more important than a relationship between the cell size and the
harp dimensions.

The number by which the column width is multiplied does need to be
controlled by the apparatus though rather than being hard coded like it
is here.