Use testing library render for snapshot testing
I don't like that the container has an extra wrapper View around the
actual component, since that stuff seems to be testing-library dependent
and could change at any minute. If I upgrade testing-library, I'll just
have to update all of the snapshots at that point, knowing that the
diffs will just be caused by that wrapper format changing.

The reason it' preferable to use testing-library is twofold:

1. I've chosen testing library over the in built renderer for various
   additional bits of functionality which may be just as useful during
snapshot testing as it is for unit testing.

2. If possible, it will be better to avoid overlapping technologies
   since that gets confusing quickly.