Create HarpFace level testResources
Making the snapshots has made it clear that the testResources which are
under the HarpFace forder of the HarpFace module are used by so many of
the siblings that it should be at the module root.

The reason it's used by so many siblings is that we're using the same
HarpFaceProps type as the parameter for as many of the components as
possible, to minimise the number of types involved in the module.