Add snapshots for different row stylings
The conclusion in the previous commit is partially correct, but it will
be helpful to have the full coverage of the HarpFace snapshot in
addition to the more targetted test here. If the styles are deliberately
changed, these tests will help identify that only the targetted
permutation has changed, while if they are changed accidentally the
HarpFace snapshot will ensure that the change is noticed even if it only
effects one of the untested rows here.