Simplify the column generation code
There was some difficulty with running the `map` function due to the
union type. You'll notice that the idSequence function is cast to a
slightly different union type to the one which is returned from the
sequence generation function. This is to deal with a type quirk which is
discussed here:
https://stackoverflow.com/questions/49510832/typescript-how-to-map-over-union-array-type