Add counterpart degree functionality to analyser
The purpose of this is to get the id required for recovering color
information. It's not relevant at the moment for the degree legend key
to know what it's counterpart pitch id is.

I don't really like the design here though. Calling something a
counterpart when it's returning the input might be confusing.