Make pitch legend keys highlight only font colour
The idea here is that when the degrees or pitches in the harpface are
displayed then their counterpart in the legend will be displayed with
the background color because then the legend key style represents the
cell in the harpface. At the moment the legend doesn't have information
on the display mode so I'm just hard wiring it to appear as though
Degrees are displayed.