Remove the display mode toggler buttons
They are a less attractive way of achieving what we already have in the
activity legend now.

I'm not committing to removing the component from the Controls module
since this replacement is only a POC at the moment and if it starts
failing I'd like the option to bring back a realiable component.