Add type to HeadsupDisplay to receive children
The functionality we are replacing in HomeScreen doesn't actually handle
children already. This is probably one step ahead of what I should be
doing before migrating the functionality, but now that I've done it I
don't think going back is worth the time. I judge this to be low risk.