Make device width dynamic in case of rotation
The app isn't always loaded in landscape mode initially. Sometimes it
seems to load in portrate and then quickly rotate. The problem with that
was that it was resulting in the device width being set here to the thin
dimension of the screen rather than the long one and not being able to
be reset after rotation.