Turn sweeping overlay in to full page object
I think I want this to actually be an object which is swept in, and then
gives the user some scrollable controls.

This will turn the component from a display to a control which wasn't my
intention. I had intended for it to just be display so that we could
swipe the harp face and see the cells move. Having seen what it looks
like with a transparent background, it's now clear that we can offer the
control through the overlay while still being able to see the harp face
well enough that the context is not lost.