Make use of nudge helper to move pozition both ways
I've had to use END state rather than ACTIVE state now since ACTIVE
doesn't have any translationY when it starts.

This means that the activeOffsetY is actually only useful for
determining whether to get to the ACTIVE state. It seems like it can get
to END without passing through ACTIVE, and there's no good way to detect
whether that is what's happened. Tap handlers have an `oldState` which
would be useful but doesn't appear to exist for Pan handlers.