Use partial application to make params generic
The one thing we know about the function which will be used to nudge the
harpstrata in one direction or the other is that it will need to
progress a list forward or backward. If the parent can apply the rest of
the function parameters up front then this nudgeFunction can be generic
in here, and can take options which progress the harpstrata in any
particular way.