Add bad function to test activation
It's not great that we've got to calculate the activation threshold
again to know whether this particular END state follows an activation or
not. But it will do for the moment. I have ideas about how to improve
which will be best explored once some more things are in place.