Use previousState to determine whether gesture activated
This work also highlighted that it's better to set your state and then
perform the logic on that state in the following render so that all the
synchronous stuff has come back together. Comparing state in the event
handler meant that my previousState value was the previous previousState
(ie 2 states back). I suspect that a lesson like this will be important
when trying to bring in animations.