Add DisplayMode nudge function
This isn't complete yet. And it's clearly having to bend around the fact
that the OptionContainer expects a nudge function which is based on the
return of a HarpStrata.

I am returning a copy of the input harpstrata because I want to make
sure that a rerender is triggered. It probably isn't necessary since
we're actually calling a state setting hook in `setActiveDisplayMode`,
but I'm just being careful for now. It can be reset to just return the
input later when I know it works.