Add DisplayMode to the CovariantMenu
I think it does make functional sense for it to be in this menu.

It
makes me realise that we need to make the various nudge functions
available at a sibling level though so that the options can be moved
around easily.

You can see that I have had to import the display nudge function in a
rather ugly way.