Add specific types to the menu options
As Pick's of the MenuProps because those are the only props that they
will have available to be passed in, and they may not need them all.

It's semantically helpful, because it helps communicate that the options
themselves are not visible beyond the module. They are only children of
the menu's and can therefore only receive props from them.