Use generic partial application to menu option instance
I have had to pull back the generic typing in the
partiallyApplyNudgeFunction function in order to get it to work.
Obviously that means it was wrong before, but I plan to make a move back
in that direction whilst being sensitive to to issue that was being
produced in the harpKeyOptionContainerProps.nudgeFunction.

I have changed the type of `setActiveHarpStrata` in the `MenuProps`
definition to be a named type. This probably wasn't part of the
solution, but since it seems quite fragile I wasn't willing to undo that
change or commit without it just in case it was important.