Export helper function so imports don't need to penetrate module
This isn't strictly the best solution. The helper should probably be at
a common level to all of the resources importing it, but it's unclear at
this stage which of the resources will remain. So it's not clear where
the best place for it is. This proper export is enough for now.