Add temp solution to orientation dimension change
This solution is simple and incomplete. It attempts to account somewhat
for the fact that when the app initially loads, it is oriented in
portrait, so the window width which is discovered and used to render the
menu offscreen turns out to only be half as far as required after the
app reorientates.

Even this approach doesn't work very well though as you can just see the
edge of the menu at the side of the screen still.

There should be a way to force a rerender when the orientation changes.