Remove unnecessary test
This test is failing now since the values displayed in the HarpCells are
split. The flat notes would need some special discovery method.

This test is somewhat redundant now that the snapshot is in place
anyway. It's still easier to read than the snapshot, but it's going to
be too hard to get it to work again so I'll chuck it out.