Prepare to turn HarpFace in to a fragment component
This work attempts to make it possible to pass a rage object to the
HarpFace construction process, thus limiting the width of the fragment
rendered.

I have not updated any tests because I'm still just proving a concept. I
may regret approaching it like this if it becomes buggy and I can't see
how to fix it.

The plan is to prove this simple concept and then refactor properly,
driven by tests afterwards.