Make HarpCellProps independent from HarpFaceProps
This is the first move in making the HarpFace related properties be
bottom up derived rather than top down. This is worth doing whilst
refining the HarpFaceFragment logic we've just proven becuase it is a
better, more flexible and semantically clearer design which will aid
in progress of the refactor.