Differentiate face and fragment level facts
Allow the HarpFaceFragment access to the face level facts if it needs
them. In this case, it needs to know how many rows there are in the
entire face, because it will want to make each fragment the same height.
However, if we were displaying these stacked vertically, we wouldn't
need them to be the same height anymore and so the fragment fact getter
should be providing a refined version of the row count in that
situation.

We're not there yet though so I've just removed the rowCount property.