Remove the Controls module
Everything remaining in here is redundant now. It served it's purpose
while I was understanding which direction to take the app. Certain
interactions needed to be provided in any form possible just to be able
to test the other parts of the app which were being developed with more
consideration.