Move the HarpFaceProps to a higher folder
I need to move `getHarpFaceFacts` up from the `HarpFace` folder since it
is required by the sytles of `HarpFaceFragment`. However, this function
uses `HarpFaceProps` as it's argument type, which would result in a
cyclic require, given that `HarpFace` requires `HarpFaceFragment` at
some point.

It seems like the idea of trying to make everything a sibling folder
inside a module is flawed in this respect and I suspect that I will be
best off just moving everything from `..HarpFace/HarpFace/*` up a level in
to `..HarpFace/*`.