Move the HarpGuru types to the module root
This has removed all of the cyclic requires warnings. It seems to be
hazardous to represent a parent module as a sibling of all of it's
dependencies. I still feel like it's a good design in many ways, but I
want to be able to rely on such warnings to mean that I am likely to be
making a mistake, and with this strategy it seems like this type of
warning is likely to result from standard practice.

I'll make sure that HarpGuru and HarpFace operate with a strictly
module-root definition strategy, like what I've moved `types` to in this
commit, and then gradually move the other modules in that direction as I
come across them in upcoming pull requests.