Move Coord type definition to HarpCell
This is preferable since this is the furthest point down the type
definition chain. We are aiming to define everything at it's most
fundamental location.

This doesn't sit quite right with me since discovering that I may be
risking cyclic requires with this kind of definition. Currently that
isn't happening and I have opened
https://github.com/js-jslog/harpnative/issues/39 to investigate a more
well researched project layout. There may be things to learn before that
is applicable so I will finish the work here before I embark on things
like that.