Introduce reactn global state and better timer
The global state means that we'll have an easier time modifying it from
the 3 places which need to update the state of this display:

1. The display itself to flip in and out of visibility
2. The cells to prompt the display to re-present itself after user
interaction
3. The other menu's to promt the display to re-present when they are
cancelled

The timer tweek here just very slightly improves to display to have a
longer gap inbetween displays than the display time itself.

The timer implementation is just a stop gap before this and all the
logic described above is put in to some global state reducers.