Use a global reducer, again just to prove that I can
It seems like the problem I've been encountering all day is just a
failure to understand and follow the rules of hooks. I understood that
the hooks couldn't be called outside the function component body, but I
didn't realise they needed to be called in the same order every time
which rules out the possibility of actually calling the useDispatch hook
inside the event handler.

I withdraw my tepid and hedged attack on Mr Stover in my previous
commit.