Add another slightly different reducer
This one just means that the question screen will be displayed
immediately rather than in 1 second from now.

Having semantically different reducers which are functionally very
similar is fine. The structure of this thing is starting to show it's
rigidity now. I might consider this branch a proof of concept if I have
to do too much more exploration without supporting tests.