Import global state file and make sure state is set
Having the initial global state set in a separate file which is imported
is a good way to ensure that it is not loaded multiple times. Node will
ensure that this is not done, even if it is imported in multiple files
for some reason.

This is not a custom hook though so it doesn't belong in this hooks folder.