Reimplement HarpCell to use a global state
I have had to mock useGlobal in all of the tests which end up rendering
HarpCell which is quite a few. These mocks aren't ideal I think. The
syntax could be better which I will work on next. I was lead around by
the nose due to the linting error which still persists, so as soon as I
got something which worked I have just stuck with it until the tests
pass.

The other thing about this step is that the app will be broken now. The
way in which it's broken should be very simple though, and I expect that
once all of the components have been reviewed like this and the global
state is set at the top of the app it will all fit together.