Remove the HarpStrata props from HarpCell props
Stating that this is now in principle independent of the props drilling
it formally used.

All of the dependencies in this module need to have their props
reviewed. I don't think it's wise or necessary to have then reference
HarpCellProps now that there's no direct drilling through from the
parent.

This and similar refactoring will be left for a followup branch though.