Remove HarpStrata props from HarpFaceFragment
Note the TODO in HarpFaceFragment styles. None of my dev tools were
telling me this was a problem and the tests passed, but I would rather
be cautious at this point until I actually see a running app. If it
turns out that the useGlobal 'hook' can be used outside of component
bodies then I will consider making wider use of that fact.