Set the global state at app root
This is an opportunity to test whether the changes made so far have tied
together to give us a functioning app. They have which is a relief and a
testement to the power of test driven refactoring.

The menu's still don't use the global state yet so that will be the next
thing to improve.