Remove HarpStrata props from all the Menus module
There's quite a lot changed here, which might be annoying if you are
analysing a single file change at this commit, but all of the changes
are very similar and they are all for the same reason; making use of the
global HarpStrata state rather than using the drilling props.

From a risk perspective this is fairly well justified as well by the
fact that I have just gone through an almost identical process in the
HarpFace module, but in a much more careful way. If you want to
understand the process then perhaps it will be useful to look back a few
commits to where this process began.