Move SetActiveHarpStrata type def to global state def
We may not even need this in the future if the number of
`setActiveHarpStrata` functions becomes sufficiently low due to the use
of `useGlobal`. For the time being it makes sense to tidy it away from
the rest of the app.

This does make me think that we have the naming all wrong for the file
which is setting the global state. I'll need to think on this.