Have reducer reset harpstrata to inactive
One very visible problem here is that the update of the global state
does not seem to produce a rerender of the cell component, so it doesn't
actually rerender as inactive until the next rerendering producing thing
occurs.

I'm unable to lead with TDD here because I don't really understand what
I'm doing yet.