Correct the harpstrata deactivation
I was completely wrong in my previous commit; the reason for the cells
not being updated was because I was hacking the reset of the harpstrata.
The cells don't use the activeDegreeIds or activePitchIds properties,
they use the isActiveMatrix, which I wasn't resetting. It's always
better to make sure you have a coherent harpstrata by calling the
harpstrata library.

It seems like the issue of failing rerendering might still be waiting
for me in the grass. I'm not certain that it is the updated global
context which is producing the rerender. If things start to fail once
the other properties are put in to global state then this article might
be of use:
https://allanbogh.com/2020/02/02/using-react-hooks-and-global-state-with-reactn/