Use reducer chain for clarity
I'm in a position now to come up with a better solution to the timing of
the quiz question presentation than to use a global counter. Here I'm
taking the opportunity to avoid using the counter state property in one
of the reducers and simultaneously do a better job of conveying that
what is actually happening in the 'quizAnswerGiven' reducer is that we
are preparing to call the 'requestNextQuestion' in a second from now.