Add a quiz question generator
I would have preferred to start doing this with DegreIds, but the
harpstrata library doesn't expose a `getDegreeIds` function yet..

Using a modified PitchIds array rather than using this recursion method
would be more efficient in theory, but this is quick and expressive
while I'm not aiming for optimisation. I'm proving a concept still.