Add an experience mode property to global state
More technical debt introduced here as the new enum doesn't belong in
the setGlobalReactNState module, but it seemed to fail when I put it in
the HarpGuru component. Perhaps I was doing something wrong, but I'll
save that investigation for when I have enough of this logic in place to
think about how to organise it.