Set global state in a function rather than in file body
I have discovered that a number of the unit tests are dependent on this
global state which is not right. The test is not loading the file
explicitly and I don't want them to be dependent on non-mocked global
states because that creates a dependency which is not only fragile, but
also very very difficult to see.

The reason this is happening appears to be that because this file is
exported by the `helpers` module, any file which is part of the test
which imports anything else from the `helpers` module will result in the
global state being set up.

I would like to keep the global state setup as it's own file to keep it
isolated, but this means that we need to either:
- have it outside of any export paths, which seems liable to regression
when I forget about this in the future
- or the actual state setup needs to be within a function which needs to
be called explicitly, which is safer as long there are measures in place
to make sure that attempts to run in more than once will fail

Currently this change breaks a load of tests and the app itself but I
thought it was valuable to make the change and capture each fix step as
a separate commit. It would be quite a confusing commit otherwise.