Use the experience mode to determine presentational facts
This is drawing close to the behaviour that I'm looking for, but I feel
very unprotected with respect to what comes next.

I want to start putting logic in to the presentation of the HarpCells,
but I feel like some of the logic there should really be inside
reducers rather than in the functions which are being used at the moment.

I've made quite a lot of untested progress, and my intention was to tidy
up and fill in the testing once I've proved the concept. However taking
the next step in proving the concept feels very hazardous without the
testing having been added.

I am wondering whether I should consider the concept sufficiently proven
that I can spend some time tidying up now, before I get to the end of
creating the this quiz mode.