Add function to guarantee that a cell will be activated
This funciton is required in order to highlight the correct answer to a
quiz question. We don't want to use a toggle function because if the
correct answer is already highlighted then we will be deactivating it
with this action. It's easier to just ensure that it's on than to do a
conditional toggling logic.

This function will probably need to move since it is likely to be useful
elsewhere when quiz questions timeout without an answer.