Correct oversight in test
This test wasn't actually testing the return value of the function at
all, it was just testing the value of the function.. which means it was
not doing anything for us.

I suppose I didn't check for a red before fixing to green, which is not
a mistake I often make - promise.