Replace drilled activeDisplayMode with global state
Just a full on replacement with the necessary consequences followed
through on with help from the linter, type checker and unit tests.

This could have been broken down in to smaller tasks as I felt was
necessary with the introduction of other global states (see recent pull
requests), however I'm a lot more confident second or third time through
and the time saved was worth the risk of time lost having to start again
and be patient.

The impact on change auditing is probably improved since all of the
changes here *really* do relate to the same underlying motivation. There
are more tidying steps to complete, but those were outside of the scope
of this functional change.

It all seems to work still so a big score for developer tooling and tests.