Redesign dispatching to be cancelled in race
There are multiple bits of logic which are going to kick off a timer for
the next question to be requested. Once any of them reach their
conclusion, all of the timers need to be cancelled so that we have the
impression of a single responsive timer. We don't want to see a flurry
of new question requests equal to the number of things which are making
new question requests.

It seems to get the HarpCell timeout to be less than around a second. I
suspect that has something to do with there being so many HarpCells
which rerender and there potentially being multiple rerenders on the
HarpCell in question which might result in the `useEffect` not being
called immediately. This is just a guess though and I'd suggest that
it's not completely in keeping with the observations. You'd expect that
under this theory, if the timeout set to `1` gave a 1 second delay, a
timeout set to `1000` would give a 2 second delay, but they both seem to
be 1 second. `2000` gives 2 seconds just fine though.

I've also had to put a TODO in for a strange error which only occurs
during the unit test run. The TODO itself explains the detail of what's
happening here.