Hookify the event handler for HarpCell long press
I've added some functionality to exit early if the pitch id is undefined
but I didn't add a test for it.

There wasn't even a test file for the previous commit.

To be honest, time is becoming tight at this point and while I am
certainly not covering every action I take with tests I feel like this
and other instances are justified and measured risks. I have other
tooling to prevent issues and the functionality I'm skipping over is
basic or already validated as working well.