Improve readibility of quiz cycle logic
The name of the `deadlineReached` function in particular says more about
what it means for that condition to be true in that context than what
the conditional logic itself is doing.

This may be considered misleading by some, but I think it makes the
intention of the `useEffect` logic *much* clearer which is the most
important thing. The reader needs much more help understanding what the
behaviour of the `useEffect` is than in interpretting what the simple
condition statement says.