Move getHarpRows to be a util of HarpFaceFragment
Previously this was it's own component folder, because the function does
wrap the HarpRow components in some more styled View elements.

Despite that and the fact that the function consequentially needs to be
a .tsx file, it is best to see this function as a util which assembles a
collection of HarpRow components.