Move OptionContainer to Stover layout
This is the first in a series of commits to see whether the structure of
the options and menus can be simplified.

While changing the file layout I have thought that the duplication of
code visible in the various specialisations of options components is
probably wasted code. It bloats the project directory making it harder
to detect the salient parts of the structure.

It might be better to just have a Option component and a Menu component
which are each simply passed some variables when it's time to build
them.

This is just an experiment and will be merged in to it's parent branch
if it is successful.