Remove need for duplicated code in option instances
Each of the options is just a variation on the params passed to the
container component. It seemed nice to isolate the construction of those
different params away so that all we were looking at was different
components, but in the end it created a lot of additional boilerplate in
the creation of those components and pointless tests testing the same
thing over and over again.

Defining the option params while building the menu still fits nicely in
to a compact component and it keeps more related code in one place.

We will see in the next few commits that the boilerplate can be reduced
even further by turning the nudge functions in to hooks.