Use hooks to perform partial application
Using hooks here alows us to move the partial application logic out of
the component body but it means we've got an unnecessary split in
functionality and tests which are complete duplicates of one another
between the hooks and utils folders.

We should try to bring the nudge functions logic in to the hooks file so
that all the logic is in one place.