Move remaining option implementations to inline definitions
Similar to previous commits. The goal here was to reduce the number of
similar components.

I think I'm happy having the various nudge functions being defined
locally to the components which use them, but obviously the display mode
one is going to need to be brought up to a common location since both of
the menu's use it.