Replace Theme folder with styles folder
We have shared styles here in the same way that we have shared versions
of every other resource, like utils and hooks.

The idea of a theme in contrast to styles is appealing but it needn't be
split in to an alternate concept. That just adds bloat to the namespace
without much benefit. The ability to swap out one colourscheme for
another is still available through the use of the multiple *-colorscheme
folders.

There could be more type safety here I defined the type of the object
which is expected to be returned from each of the colourscheme folders.