Simplify props
I have discovered that when all the props are named `Props` some of the
type warning messages can be confusing.

I may have gone too far with the utils functions in harp-rows, but we'll
see. I'm only trying to be consistent and it seems like object params
are unnecessary when there are fewer than 3 params.