Add overhang label tab to menu
The menu's need some way of making their presence known while tucked
away. Having a label poking out from the side from which they should be
swiped in will achieve this.

The formation of this POC is not at all ideal. The MenuContainer and
AnimatedMenuContainer concepts are not really working for me now so I
will rework these for the CovariantMenu (soon to be TuningMenu)
specifically before turning towards making it a bit more generic.

I'll need to consider the fact that not all the overlays will need this
lip. The quiz question display requires all the same transparency
styling but without a lip. We don't want it's presence to be advertised
at all.