Add animation and label directly to covariant menu
The animation was copied from the work done in
https://github.com/js-jslog/harpnative/issues/56#issuecomment-667687204

The reason I've put all of the logic inside the menu here is because I
can repeat the process for the other menu as well as the quiz display
and then refactor out the repeated code. That'll result in a better
design than trying to optimise early.