Apply similar but restricted technique to question display
I've made some non-functional changes to styling of this component.
Meaning the structure is different but the result is the same. The
purpose of this is to bring this overlay component as close to the
styling of the menus as possible so that similarities can be seen as
easily as possible. From here their commonalities can be captured as
shared code.