Fix zindex so that overlays are in front of number row
The number row was previously appearing in front of the menu's and quiz
popup and with the transparency effect in place it was very obvious that
this row was popping through.

I've left some indexes below and above the hole number row so that we
don't necessarily need to shuffle them round if more indexed layers are
added.