Remove harp row boarder for decluttering
This definitely reduces the feeling of clutter but still achieves the
home row identification and fragment highlighting if a bit more subtly.

I have left the boarder in at a width of 0 since it would be there
implicitely anyway and this way it's easy to play with in the future.