Add experimental scaling animation for menu
I like the 'tab' effect that the label has like this.

I can't decide whether I like the way the boardered
menu scales in to view.

I definitely don't like the way the label font grows to be enormous
during it's translation. I think it either needs to be scaled back while
the rest of it grows or made gradually transparent.