Update snapshots with recent changes
Again, these should have been updated with a previous commit when the
functionality was changed.

I could consider using husky to prevent these mistakes.

The reason is that I have left a TDD methodology as I start tinkering
with the ux and design. I have no way to reliably define tests ahead of
creating this functionality.