Make menu label a transitioning colour
This works well to call attention to the label tab without having a
horrible line moving across the page as it expands.

The naming isn't brilliant anymore. We've got a lot of animation values
and not much specificity in where they are being used.

I've also been unable to get the background colour working on the
existing Animated.View, so I've had to add another one which isn't so
tidy.

Anyway, it works and it can be improved on.