Remove all hooks from harp-cells
This has indeed reduced the number of renders significantly and sped
everything up a lot. However, it does now mean that the toggleHarpCell
and setPozitionRoot callbacks are no longer accurate since they need to
know the active cells which are updated at every interaction. This might
mean that each cell *has* to be rerendered after each interaction or
else it's callback will be outdated.

If the callback could somehow be linked to some mutable but indisposable
reference which itself maintained the list of active cells then perhaps
these cells could get away without needing the callbacks to be refreshed
constantly.