Revert attempts to animate cell colouring
This reverts commits:

- 74b0464c2536baf767234744efdc946ac7af0a8d.
- 69dbcd3c7fa4613875cea8373e2b7daa51c78326
- 74b0464c2536baf767234744efdc946ac7af0a8d

If this could work then it would be a great effect and well worth doing.

It's not worth doing at the expense of performance though and while I
don't have a satisfactory solution, I'll leave it as something to come
back to.

I've reverted so that we have a history of these commits for later
reference.